np2 copy False -> None

+++ src/aspire/abinitio/commonline_base.py
@@ -149,7 +149,7 @@ class CLOrient3D:
             #   Apply mask in doubles (allow imgs to upcast as needed)
             imgs = imgs * fuzz_mask
             #   Cast to desired type
-            imgs = Image(imgs.asnumpy().astype(self.dtype, copy=False))
+            imgs = Image(imgs.asnumpy().astype(self.dtype, copy=None))
 
         # Obtain coefficients of polar Fourier transform for input 2D images
         pft = PolarFT(
